+++ pyannote/audio/pipelines/clustering.py
@@ -386,7 +386,8 @@ class AgglomerativeClustering(BaseClustering):
         elif num_large_clusters > max_clusters:
             num_clusters = max_clusters
 
-        if num_clusters is not None:
+        # look for perfect candidate if necessary
+        if num_clusters is not None and num_large_clusters != num_clusters:
             # switch stopping criterion from "inter-cluster distance" stopping to "iteration index"
             _dendrogram = np.copy(dendrogram)
             _dendrogram[:, 2] = np.arange(num_embeddings - 1)