+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix(pipeline): add missing "embedding" hook call in `SpeakerDiarization`
 - feat(utils): add `"soft"` option to `Powerset.to_multilabel`
 - improve(pipeline): compute `fbank` on GPU when requested
+- fix(pipeline): fix `AgglomerativeClustering` to honor `num_clusters` when provided
 
 ## Version 3.0.1 (2023-09-28)